more auth, cleaning up controllers, and flash messages

+++ db/migrate/20140919220127_create_events.rb
@@ -6,8 +6,8 @@ class CreateEvents < ActiveRecord::Migration
       t.string :hashtag
       t.string :desc
       t.string :loc
-      t.integer :lat
-      t.integer :long
+      t.float :lat
+      t.float :long
       t.string :tweet_count
       t.string :date
